requirements.txt
uvicorn
fastapi
jinja2
python-multipart
bcrypt
pyJWT
python-dotenv
httpx